requirements.txt
replicate>=0.25.0
streamlit>=1.28.0
Pillow>=10.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
minio>=7.2.0
python-dotenv>=1.0.0

